// src/types/BDI.ts
export interface BDI{
    id: number;
    administracaoCentral: number;
    administracaoLocal: number;
    despesasIndiretas: number;
    despesaFinanceira: number;
    posObras: number;
    risco: number;
    impostos: number;
    margemLucro: number;
    seguro: number;
    reservaTecnica: number;
    obraId: number;
}